fix: moving extra components out of toolbar
