+++ packages/strapi-plugin-rich-text/admin/src/components/RichText/Icons/Youtube.tsx
@@ -1,0 +1,15 @@
+const Youtube = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="1rem"
+    height="1rem"
+    viewBox="0 0 24 24"
+    fill="none"
+    className="extra-icon"
+  >
+    <path d="M2.5 17a24.12 24.12 0 0 1 0-10 2 2 0 0 1 1.4-1.4 49.56 49.56 0 0 1 16.2 0A2 2 0 0 1 21.5 7a24.12 24.12 0 0 1 0 10 2 2 0 0 1-1.4 1.4 49.55 49.55 0 0 1-16.2 0A2 2 0 0 1 2.5 17" />
+    <path id="bulb" d="m10 15 5-3-5-3z" />
+  </svg>
+);
+
+export default Youtube;
