+++ packages/strapi-plugin-rich-text/admin/src/components/RichText/Toolbar.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { Editor } from "@tiptap/react";
 import { Box } from "@strapi/design-system/Box";
 import { Flex } from "@strapi/design-system/Flex";
@@ -13,15 +13,13 @@ import {
   Link,
   Minus,
 } from "@strapi/icons";
-import { Dialog, DialogBody, DialogFooter } from "@strapi/design-system/Dialog";
-import { Stack } from "@strapi/design-system/Stack";
-import { Button } from "@strapi/design-system/Button";
-import { TextInput } from "@strapi/design-system/TextInput";
-import { Select, Option } from "@strapi/design-system/Select";
-import { Checkbox } from "@strapi/design-system/Checkbox";
 
-import { StyledToolbar } from "./Toolbar.styles";
 import AbbrDialog from "./Dialogs/AbbrDialog";
+import InsertLinkDialog from "./Dialogs/InsertLinkDialog";
+import InsertYouTubeDialog from "./Dialogs/InsertYouTubeDialog";
+import { StyledToolbar } from "./Toolbar.styles";
+import BlockTypeSelect from "./Components/BlockTypeSelect";
+import Youtube from "./Icons/Youtube";
 
 interface ToolbarProps {
   editor: Editor | null;
@@ -97,7 +95,7 @@ export default function Toolbar({ editor }: ToolbarProps) {
                   }
                 />
                 <IconButton
-                  icon={ExtraIcons.YouTube}
+                  icon={<Youtube />}
                   label="YouTube"
                   className={[
                     "large-icon",
@@ -139,321 +137,3 @@ export default function Toolbar({ editor }: ToolbarProps) {
     </>
   );
 }
-
-function BlockTypeSelect({ editor }: { editor: Editor }) {
-  const [selectedType, setSelectedType] = useState<string>("paragraph");
-
-  const onSelect = useCallback((type: string) => {
-    switch (type) {
-      case "h1":
-        editor.chain().focus().toggleHeading({ level: 1 }).run();
-        break;
-      case "h2":
-        editor.chain().focus().toggleHeading({ level: 2 }).run();
-        break;
-      case "h3":
-        editor.chain().focus().toggleHeading({ level: 3 }).run();
-        break;
-      case "h4":
-        editor.chain().focus().toggleHeading({ level: 4 }).run();
-        break;
-      case "h5":
-        editor.chain().focus().toggleHeading({ level: 5 }).run();
-        break;
-      case "h6":
-        editor.chain().focus().toggleHeading({ level: 6 }).run();
-        break;
-      case "blockquote":
-        editor.chain().focus().toggleBlockquote().run();
-        break;
-      case "orderedList":
-        editor.chain().focus().toggleOrderedList().run();
-        break;
-      case "bulletList":
-        editor.chain().focus().toggleBulletList().run();
-        break;
-      case "paragraph":
-        editor.chain().focus().setParagraph().run();
-        break;
-    }
-
-    setTimeout(() => {
-      editor.commands.focus();
-    }, 50);
-  }, []);
-
-  const setActiveType = useCallback(() => {
-    if (editor.isActive("heading", { level: 1 })) setSelectedType("h1");
-    if (editor.isActive("heading", { level: 2 })) setSelectedType("h2");
-    if (editor.isActive("heading", { level: 3 })) setSelectedType("h3");
-    if (editor.isActive("heading", { level: 4 })) setSelectedType("h4");
-    if (editor.isActive("heading", { level: 5 })) setSelectedType("h5");
-    if (editor.isActive("heading", { level: 6 })) setSelectedType("h6");
-    if (editor.isActive("paragraph")) setSelectedType("paragraph");
-    if (editor.isActive("blockquote")) setSelectedType("blockquote");
-    if (editor.isActive("orderedList")) setSelectedType("orderedList");
-    if (editor.isActive("bulletList")) setSelectedType("bulletList");
-  }, []);
-
-  useEffect(() => {
-    editor.on("selectionUpdate", setActiveType);
-    return () => {
-      editor.off("selectionUpdate", setActiveType);
-    };
-  }, [editor]);
-
-  return (
-    <Select
-      required
-      size="S"
-      placeholder="Text Style"
-      onChange={onSelect}
-      value={selectedType}
-    >
-      <Option value={"paragraph"}>Paragraph</Option>
-      <Option value={"h1"}>Heading 1</Option>
-      <Option value={"h2"}>Heading 2</Option>
-      <Option value={"h3"}>Heading 3</Option>
-      <Option value={"h4"}>Heading 4</Option>
-      <Option value={"blockquote"}>Quote</Option>
-      <Option value={"orderedList"}>Ordered list</Option>
-      <Option value={"bulletList"}>Bullet list</Option>
-    </Select>
-  );
-}
-
-type DialogProps = {
-  editor: Editor;
-  onExit: () => void;
-};
-
-function InsertLinkDialog({ editor, onExit }: DialogProps) {
-  const [href, setHref] = useState<string>("");
-  const [newTab, setNewTab] = useState<boolean>(false);
-  const [shouldRemove, setShouldRemove] = useState<boolean>(false);
-
-  const onClose = useCallback(() => {
-    setHref("");
-    setNewTab(false);
-    onExit();
-  }, []);
-
-  const onInsertLink = useCallback(() => {
-    if (!href || shouldRemove) {
-      editor.chain().focus().extendMarkRange("link").unsetLink().run();
-    } else {
-      editor
-        .chain()
-        .focus()
-        .extendMarkRange("link")
-        .setLink({ href, target: newTab ? "_blank" : "_self" })
-        .run();
-    }
-
-    onClose();
-  }, [editor, href]);
-
-  useEffect(() => {
-    if (editor.isActive("link")) {
-      const { href, target } = editor.getAttributes("link") as {
-        href: string;
-        target: string;
-      };
-      setHref(href);
-      setNewTab(target === "_blank");
-      setShouldRemove(true);
-    }
-
-    return () => {
-      setHref("");
-      setNewTab(false);
-      setShouldRemove(false);
-    };
-  }, []);
-
-  return (
-    <Dialog onClose={onClose} title="Insert link" isOpen={true}>
-      <DialogBody>
-        <Stack spacing={2}>
-          <TextInput
-            label="Link URL"
-            placeholder="Write or paste the url here"
-            name="url"
-            onChange={(e: ChangeEvent<HTMLInputElement>) => {
-              setHref(e.target.value);
-              setShouldRemove(false);
-            }}
-            value={href}
-            aria-label="URL"
-          />
-          <Checkbox
-            value={newTab}
-            onValueChange={(v: boolean) => {
-              setNewTab(v);
-              setShouldRemove(false);
-            }}
-          >
-            Open in new tab
-          </Checkbox>
-        </Stack>
-      </DialogBody>
-      <DialogFooter
-        startAction={
-          <Button onClick={onClose} variant="tertiary">
-            Cancel
-          </Button>
-        }
-        endAction={
-          <Button
-            onClick={() => onInsertLink()}
-            variant={shouldRemove ? "danger-light" : "success-light"}
-          >
-            {shouldRemove ? "Remove" : "Insert"} Link
-          </Button>
-        }
-      />
-    </Dialog>
-  );
-}
-
-function InsertYouTubeDialog({ editor, onExit }: DialogProps) {
-  const [src, setSrc] = useState("");
-  const [fixedDimensions, setFixedDimensions] = useState(false);
-  const [height, setHeight] = useState<number | string>(480);
-  const [width, setWidth] = useState<number | string>(640);
-
-  const onInsert = useCallback(
-    ({
-      src,
-      height,
-      width,
-      fixedDimensions,
-    }: {
-      src: string;
-      height: number | string;
-      width: number | string;
-      fixedDimensions: boolean;
-    }) => {
-      try {
-        editor
-          .chain()
-          .focus()
-          .setYoutubeVideo({
-            src,
-            width: fixedDimensions
-              ? typeof width === "number"
-                ? width
-                : parseInt(width, 10)
-              : undefined,
-            height: fixedDimensions
-              ? typeof height === "number"
-                ? height
-                : parseInt(height, 10)
-              : undefined,
-          })
-          .run();
-        onExit();
-      } catch (error) {
-        console.error(error);
-      }
-    },
-    [editor, onExit]
-  );
-
-  return (
-    <Dialog onClose={onExit} title="Insert YouTube embed" isOpen={true}>
-      <DialogBody>
-        <Stack spacing={2}>
-          <TextInput
-            label="YouTube URL"
-            placeholder="Add YouTube URL"
-            name="url"
-            value={src}
-            onChange={(e: ChangeEvent<HTMLInputElement>) =>
-              setSrc(e.target.value)
-            }
-            aria-label="YouTube URL"
-          />
-          <Box style={{ marginTop: "20px" }}>
-            <Checkbox
-              value={fixedDimensions}
-              onValueChange={(v: boolean) => {
-                setFixedDimensions(v);
-              }}
-            >
-              Set Fixed Dimensions
-            </Checkbox>
-          </Box>
-          <Stack
-            horizontal={true}
-            spacing={2}
-            style={
-              !fixedDimensions
-                ? {
-                    pointerEvents: "none",
-                    opacity: 0.5,
-                    filter: "grayscale(1)",
-                  }
-                : undefined
-            }
-          >
-            <TextInput
-              label="Width"
-              type="number"
-              placeholder="Add Width"
-              name="width"
-              value={width}
-              onChange={(e: ChangeEvent<HTMLInputElement>) =>
-                setWidth(e.target.value)
-              }
-              aria-label="YouTube Video Width"
-            />
-            <TextInput
-              label="Height"
-              type="number"
-              placeholder="Add Height"
-              name="height"
-              value={height}
-              onChange={(e: ChangeEvent<HTMLInputElement>) =>
-                setHeight(e.target.value)
-              }
-              aria-label="YouTube Video Height"
-            />
-          </Stack>
-        </Stack>
-      </DialogBody>
-      <DialogFooter
-        startAction={
-          <Button onClick={onExit} variant="tertiary">
-            Cancel
-          </Button>
-        }
-        endAction={
-          <Button
-            disabled={src.length === 0}
-            onClick={() => onInsert({ src, width, height, fixedDimensions })}
-            variant="success-light"
-          >
-            Insert YouTube Embed
-          </Button>
-        }
-      />
-    </Dialog>
-  );
-}
-
-const ExtraIcons = {
-  YouTube: (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="1rem"
-      height="1rem"
-      viewBox="0 0 24 24"
-      fill="none"
-      className="extra-icon"
-    >
-      <path d="M2.5 17a24.12 24.12 0 0 1 0-10 2 2 0 0 1 1.4-1.4 49.56 49.56 0 0 1 16.2 0A2 2 0 0 1 21.5 7a24.12 24.12 0 0 1 0 10 2 2 0 0 1-1.4 1.4 49.55 49.55 0 0 1-16.2 0A2 2 0 0 1 2.5 17" />
-      <path id="bulb" d="m10 15 5-3-5-3z" />
-    </svg>
-  ),
-};
